Use pwd as project path if run from repo
Before the pdb and extensions paths were hard coded, now it will detect
if you are running from the root of a pdb or extensions repo and use
that path instead. The other path will remain the default. If run from
a directory that is not the root of a pdb or extenstions repo both
defaults will be used.

This has simplified testing PRs that make changes to benchmark and the
pdbbox scripts.